+++ src/integrators/HeadlessIntegrator.php
@@ -150,6 +150,22 @@ class HeadlessIntegrator extends BaseObject implements IntegratorInterface
             ->isSuccess();
     }
 
+    /**
+     * {@inheritDoc}
+     */
+    public function getProviderData(PayModel $model): array
+    {
+        $model = ApiPaymentProcess::view($model->id)
+            ->setFields(['id', 'provider_data'])
+            ->one($this->getClient());
+
+        if (!$model) {
+            throw new PaymentException("Unable to find the given payment model.");
+        }
+
+        return $model->provider_data;
+    }
+
     // internal
 
     private static function createPayModel(ApiPaymentProcess $process)