+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 In order to read more about upgrading and BC breaks have a look at the [UPGRADE Document](UPGRADE.md).
 
+## 3.0
+
+> This release contains significant changes regarding SaferPay integration. Check the [UPGRADE document](UPGRADE.md) to read more about breaking changes.
+
++ [#22](https://github.com/luyadev/luya-module-payment/pull/22) Remove SaferPay HTTPS Interface provider as its deprecated until end of 2020.
+
 ## 2.1 (11. October 2020)
 
 + [#21](https://github.com/luyadev/luya-module-payment/pull/21) New methods to close a payment process within a transaction.